MDEV-18465 Logging of DDL statements during backup
Many of the changes was needed to be able to collect and print engine
name and table version id's in the ddl log.

Other things:
- Backup START now waits for running or waiting DDL's to complete
(backup_lock_debug.test is changed to take this into account)
- Added test that CREATE VIEW is blocked by BLOCK_DDL
- ha_resolve_storage_engine_name() now also works for views

-------------------------------------------------------------------------------
MENT-222: Server crashes in add_str_to_buffer / backup_log_ddl upon CREATE TABLE IF NOT EXISTS .. LIKE

At the end of mysql_create_like_table() function DDL log info is filled and
logged. Some fields of DDL log info are filled from local_create_info local
variable. If table exists, then local_create_info stays unitialised at the
moment of DDL log info filling.

The fix is to fill and print DDL log info only if table does not exist.

-------------------------------------------------------------------------------
MENT-482: main.backup_log fails with --ps-protocol mtr option

Table opening order is different for mysql_create_like_table() and
mysql_test_create_table()(which is invoked from check_prepared_statement()).

That is why when "CREATE TABLE IF NOT EXISTS t_exists LIKE v" is
executed, 't_exists' table is opened firstly, and, as it exists, the
statement execution returns 'ok' and does not create table.

But when the statement is prepared, 'v' table is opened firstly, and as
it does not exist, the error occured.

The fix is to have existing table in 'LIKE' part of the statement to
avoid error when ps protocol is used for mtr testing.

-------------------------------------------------------------------------------
MENT-368 Race condition in backup ddl logging

The race condition allowed a call to log a ddl query while the log file
was getting closed.
This could potentially write to a file descriptor that was reused by
some other thread.