MENT-679: wsrep_plugin and InnoDB plugin initialization order changed with S3 plugin
Plugin initialization order is stored in a hash and only limited amount
of bytes are used from plugin name. When S3 was introduced some
compilation modes this could cause order of wsrep_plugin and InnoDB
plugin initialization to change. Fix is to use MariaDB_FUNCTION_PLUGIN
for wsrep_plugin to make sure that wsrep_plugin is initialized
before InnoDB plugin.