MENT-830 Fix InnoDB log resizing
The problem:
- first thread takes fd
- second thread swaps the same fd with another and closes an old one
- first thread call fsync() on a closed fd

This is a lifetime issue. This patche solves it by extending a lifetime
of an fd for the whole time of write(fd) and fsync(fd).
Lifetime is now managed by std::shared_ptr<file_io>.
Writer thread takes a copy of that shared_ptr and owns it throught the
whole duration of write(fd) or fsync(fd).

Note that shared_ptr itself is shared. Checkpoint thread may replace it
with another one. Thus, it's protected with std::mutex.

Every write() and fsync() now take a copy of std::shared_ptr<T> under
the mutex protection. Copying is memcpy(2 * sizeof(void*)) and
fetch_add(1, std::memory_order_relaxed). It's a fast operation, a nothing
basically, comparing to the cost of syscalls.