MENT-830 ib_logresized is written to the wrong offset, breaking recovery
log_t::log::lsn_offset is a stuff specific to a redo log. Server
initializes it once on a database bootstrapping and then it's
value is incremented with wrapping around.

The same should be true for a new redo log file. It can't use the old
lsn_offset or cases when old lsn_offset > new_file_size are possible.
And this is just one the most obvious possible bug.

This patch creates an lsn_offset for a new redo log file and manages it
separately. When new file is adopted, new lsn_offset is adopted too.